+++ src/compiler/index.ts
@@ -7,7 +7,7 @@
 
 export { KipperCompiler, KipperCompileResult } from "./compiler";
 export { KipperErrorListener } from "./error-handler";
-export { KipperStreams, KipperParseFile, KipperParseStream } from "./parse-stream";
+export { KipperStreams, KipperParseStream } from "./parse-stream";
 export { KipperFileListener } from "./listener";
 export { KipperFileContext } from "./file-ctx";
 export * as parser from "./parser";